+++ midware/src/app_timer.c
@@ -10,7 +10,7 @@
 
 #define TIMER3_PCLK_DIV         (256)
 
-uint8_t blink_cnt;
+uint32_t blink_cnt;
 
 static stc_tim3_mode0_cfg_t timer3_reg_cfg;
